fix amount type

+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/services/remote/BankRestController.java
@@ -20,7 +20,7 @@ public class BankRestController {
 	private static Logger logger = LoggerFactory.getLogger(BankRestController.class);
 
 	@RequestMapping(value = "/accounts/{iban}/processPayment", method = RequestMethod.POST)
-	public ResponseEntity<String> processPayment(@PathVariable String iban, @RequestParam int amount) {
+	public ResponseEntity<String> processPayment(@PathVariable String iban, @RequestParam double amount) {
 		logger.info("processPayment iban:{}, amount:{}", iban, amount);
 		try {
 			return new ResponseEntity<>(BankInterface.processPayment(iban, amount), HttpStatus.OK);
